+++ NEWS.md
@@ -1,5 +1,7 @@
 # uisapi (development version)
 
+* Fix bug with parsing of `last_data_update` in `uis_get_indicators()` due to date format changes in the API.
+
 # uisapi 0.1.0
 
 * Initial release with `uis_get()`, `uis_get_entities()`, `uis_get_indicators()`, `uis_get_versions()`, and `uis_bulk_files()`